move Func, BoolFunc, tests as they require go1.21
Commit 69bc3bd5b37fa90e994be9acecf7430269591713 added support for Func()
and BoolFunc() to match stdlib. However, the Func method was added in
[go1.16.0], and BoolFunc in [go1.21.0], so running the tests on older
versions of Go would fail;

    docker run -it --rm -v ./:/pflag -w /pflag golang:1.21 sh -c 'go test -v ./...'
    # github.com/spf13/pflag [github.com/spf13/pflag.test]
    ./bool_func_test.go:86:28: cannot use stdFSet (type *flag.FlagSet) as type BoolFuncFlagSet in argument to runCase:
        *flag.FlagSet does not implement BoolFuncFlagSet (missing BoolFunc method)
    ./bool_func_test.go:113:21: undefined: io.Discard
    ./bool_func_test.go:116:28: cannot use stdFSet (type *flag.FlagSet) as type BoolFuncFlagSet in argument to runCase:
        *flag.FlagSet does not implement BoolFuncFlagSet (missing BoolFunc method)
    ./bool_func_test.go:139:7: undefined: errors.Is
    ./func_test.go:92:28: cannot use stdFSet (type *flag.FlagSet) as type FuncFlagSet in argument to runCase:
        *flag.FlagSet does not implement FuncFlagSet (missing Func method)
    ./func_test.go:119:21: undefined: io.Discard
    ./func_test.go:122:28: cannot use stdFSet (type *flag.FlagSet) as type FuncFlagSet in argument to runCase:
        *flag.FlagSet does not implement FuncFlagSet (missing Func method)
    ./func_test.go:145:7: undefined: errors.Is
    ./func_test.go:145:7: too many errors
    FAIL	github.com/spf13/pflag [build failed]

This patch moves the tests to a separate file that is not built for older
versions of Go.

[go1.16.0]: https://pkg.go.dev/flag@go1.16.0#Func
[go1.21.0]: https://pkg.go.dev/flag@go1.21.0#BoolFunc
